trying to fix github actions

+++ dataanalyzer/utils/run.js
@@ -7,23 +7,25 @@ module.exports = async function run() {
     // const comments = await Comment.find({}).sort({ timestamp: -1 }).limit(30).exec();
     const comments = await findComments();
     const frequencies = {};
-    let max = comments[0].coinTicker;
-    for (let comment of comments) {
-        frequencies[comment.coinTicker] = frequencies[comment.coinTicker] ? frequencies[comment.coinTicker] + 1 : 1;
-        if(frequencies[comment.coinTicker] > frequencies[max]){
-            max = comment.coinTicker;
+    if(comments != null && comments.length > 0){
+        let max = comments[0].coinTicker;
+        for (let comment of comments) {
+            frequencies[comment.coinTicker] = frequencies[comment.coinTicker] ? frequencies[comment.coinTicker] + 1 : 1;
+            if(frequencies[comment.coinTicker] > frequencies[max]){
+                max = comment.coinTicker;
+            }
         }
+    
+        await Trending.deleteMany({});
+        const trending = {
+            coinTicker: max,
+            numComments: frequencies[max]
+        }
+    
+        let dbTrending = new Trending(trending);
+        await dbTrending.save();
+    
+        console.log("Data analyzer completed.");
+        process.exit();
     }
-
-    await Trending.deleteMany({});
-    const trending = {
-        coinTicker: max,
-        numComments: frequencies[max]
-    }
-
-    let dbTrending = new Trending(trending);
-    await dbTrending.save();
-
-    console.log("Data analyzer completed.");
-    process.exit();
 }
